Rename gw to lfgw
