+++ internal/lfgw/helpers.go
@@ -1,4 +1,4 @@
-package gw
+package lfgw
 
 import (
 	"fmt"
