+++ internal/lfgw/logging.go
@@ -1,4 +1,4 @@
-package gw
+package lfgw
 
 import (
 	"log"
