+++ internal/lfgw/main_test.go
@@ -1,4 +1,4 @@
-package gw
+package lfgw
 
 import (
 	"flag"
