+++ internal/lfgw/middleware_test.go
@@ -1,4 +1,4 @@
-package gw
+package lfgw
 
 import (
 	"net/http"
