+++ internal/lfgw/routes.go
@@ -1,4 +1,4 @@
-package gw
+package lfgw
 
 import (
 	"github.com/gorilla/mux"
